# perf(core): skip await if response has already resolved