--- a/packages/yarnpkg-core/sources/httpUtils.ts
+++ b/packages/yarnpkg-core/sources/httpUtils.ts
@@ -12,7 +12,7 @@ const NETWORK_CONCURRENCY = 8;
 
 const limit = plimit(NETWORK_CONCURRENCY);
 
-const cache = new Map<string, Promise<Response<Buffer>>>();
+const cache = new Map<string, Promise<Buffer> | Buffer>();
 
 const globalHttpAgent = new HttpAgent({keepAlive: true});
 const globalHttpsAgent = new HttpsAgent({keepAlive: true});
@@ -97,16 +97,20 @@ export async function get(target: string, {configuration, json, ...rest}: Option
   let entry = cache.get(target);
 
   if (!entry) {
-    entry = request(target, null, {configuration, ...rest});
+    entry = request(target, null, {configuration, ...rest}).then(response => {
+      cache.set(target, response.body);
+      return response.body;
+    });
     cache.set(target, entry);
   }
 
-  const response = await entry;
+  if (Buffer.isBuffer(entry) === false)
+    entry = await entry;
 
   if (json) {
-    return JSON.parse(response.body.toString());
+    return JSON.parse(entry.toString());
   } else {
-    return response.body;
+    return entry;
   }
 }