**What's the problem this PR addresses?**

The resolution step is currently a bit slow and uses a decent amount of memory.

**How did you fix it?**

- Once the GET request finishes, replace the entry in the cache with the response body
- If the entry from the cache is a buffer, skip the await call.

**Results**
Average measurements over three runs with no lockfile
||Total dependencies|Time before* | Time after*| RAM before** | RAM after**|
|--|--|--|--|--|--|
|next@9.4.5-canary.3  | 784 | 1881ms | 811ms|425MB | 196MB|
|react-scripts@3.4.1|1408 |  3633ms | 1325ms |642MB | 237MB|

*Time: `Self Time` of the `get` function measured by the profiler in Chrome DevTools
**RAM: Private bytes as measured by my task manager, after using Chrome DevTools' `Collect garbage` function three times

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.